classes/swupdate-common.bbclass: Add SRC_URI exclusions
Adds a variable which can be used to exclude specific files in SRC_URI
from being placed in the CPIO archive. This can be useful if a recipe
has extra actions it takes based on SRC_URI files, but doesn't want to
have those files in the CPIO archive (e.g. expanding a sw-description
using Jinja2 templates).

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>